+++ CHANGES.md
@@ -14,6 +14,8 @@
 <!-- Changes that affect Black's stable style -->
 
 - Don't double-decode input, causing non-UTF-8 files to be corrupted (#4964)
+- Fix `--line-ranges` not marking the first block as changed when it contains
+  modifications before the first matching content (#4997)
 
 ### Preview style